1.增加接口类基类配置项 2.增加maker配置文件
